auth-service/requirements.txt
Flask==2.3.3
Flask-SQLAlchemy==3.1.1
psycopg2-binary==2.9.9
Werkzeug==2.3.7
PyJWT==2.8.0
marshmallow==3.20.1
